chore: Update auth module imports to use services folder
