+++ libs/auth/src/guards/roles.guard.ts
@@ -7,13 +7,14 @@ import {
 	ForbiddenException,
 } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
-import { AuthService } from '../auth.service';
+//import { AuthService } from '../services/auth.service';
+import { LandlordAuthService } from '../services/landlord-auth.service';
 
 @Injectable()
 export class RolesGuard implements CanActivate {
 	constructor(
 		private reflector: Reflector,
-		private authService: AuthService,
+		private authService: LandlordAuthService,
 	) {}
 
 	async canActivate(context: ExecutionContext): Promise<boolean> {
