+++ libs/auth/src/index.ts
@@ -1,5 +1,5 @@
 export * from './auth.module';
-export * from './auth.service';
+export * from './services/auth.service';
 export * from './dto/auth-response.dto';
 export * from './dto/user-login.dto';
 export * from './types/firebase.types';
