+++ libs/common/src/permissions/feature-permission.service.ts
@@ -19,7 +19,7 @@ export class FeaturePermissionService {
 	private readonly cacheService = new CacheService(this.cacheManager);
 	constructor(
 		private readonly featuresPermissionRepository: FeaturesPermissionRepository,
-		@InjectMapper() private readonly mapper: Mapper,
+		@InjectMapper('MAPPER') private readonly mapper: Mapper,
 		@Inject(CACHE_MANAGER) private cacheManager: Cache,
 	) {}
