+++ libs/common/src/services/properties-amenity.service.ts
@@ -22,7 +22,7 @@ export class PropertiesAmenityService {
 	private readonly cacheService = new CacheService(this.cacheManager);
 	constructor(
 		private readonly propertyAmenityRepository: PropertyAmenityRepository,
-		@InjectMapper() private readonly mapper: Mapper,
+		@InjectMapper('MAPPER') private readonly mapper: Mapper,
 		@Inject(CACHE_MANAGER) private cacheManager: Cache,
 	) {}
