+++ libs/common/src/services/properties-purpose.service.ts
@@ -26,7 +26,7 @@ export class PropertiesPurposeService {
 	private readonly cacheService = new CacheService(this.cacheManager);
 	constructor(
 		private readonly propertyPurposeRepository: PropertyPurposeRepository,
-		@InjectMapper() private readonly mapper: Mapper,
+		@InjectMapper('MAPPER') private readonly mapper: Mapper,
 		@Inject(CACHE_MANAGER) private cacheManager: Cache,
 	) {}
